requirements.txt
Flask==2.3.3
python-telegram-bot==20.7
groq==0.9.0
httpx==0.25.2
python-dotenv==1.0.0
